+++ src/test/kotlin/domain/InputValidatorTest.kt
@@ -1,4 +1,5 @@
-import domain.InputValidator
+package domain
+
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
